+++ src/lib.rs
@@ -2,6 +2,7 @@
 // the `docsrs` configuration attribute is defined
 #![cfg_attr(docsrs, feature(doc_auto_cfg))]
 #![doc = include_str!("../README.md")]
+extern crate core;
 
 pub use color_profile::*;
 pub use context::HeifContext;
@@ -17,7 +18,6 @@ pub use reader::{Reader, StreamReader};
 #[cfg(feature = "v1_19")]
 pub use security_limits::*;
 pub use utils::check_file_type;
-
 mod color_profile;
 mod context;
 mod decoder;
@@ -27,6 +27,8 @@ mod errors;
 mod heif;
 mod image;
 mod image_handle;
+#[cfg(feature = "image")]
+pub mod integration;
 mod metadata;
 mod reader;
 #[cfg(feature = "v1_18")]